fix json invalid number
By default on C# the float to string conversion changes based on the OS region. This can cause the float numbers to be represented with a decimal comma instead of a decimal point and result in a "Invalid number" JSON syntax error. Therefore, the conversion function has been changed to ensure that, with independence of the region, the float values are going to be written with a decimal point separator.